rejigger source/thumbnails
