+++ zuds/source.py
@@ -4,144 +4,160 @@ from sqlalchemy import func, Index
 from sqlalchemy.dialects import postgresql as psql
 from astropy.table import Table
 
-from .core import DBSession, without_database, Source
+from .core import DBSession, without_database, Base
 from .utils import fid_map
 from .image import CalibratableImage
 from .photometry import ForcedPhotometry
 from .detections import Detection
 from .subtraction import SingleEpochSubtraction
 from .constants import MJD_TO_JD
+from .thumbnails import Thumbnail
+from .spatial import SpatiallyIndexed
 
 import numpy as np
 
 __all__ = ['Source']
 
-# Extension of the skyportal default Source class
-Source.thumbnails = relationship('Thumbnail', cascade='all')
-Source.accumulated_rb = sa.Column(sa.Float)
 
+class Source(Base, SpatiallyIndexed):
+    id = sa.Column(sa.String, primary_key=True)
+    # TODO should this column type be decimal? fixed-precison numeric
+    ra = sa.Column(sa.Float)
+    dec = sa.Column(sa.Float)
 
-@without_database([])
-def images(self, type=CalibratableImage):
+    offset = sa.Column(sa.Float, default=0.0)
+    redshift = sa.Column(sa.Float, nullable=True)
+    neighbor_info = sa.Column(psql.JSONB)
 
-    candidates = DBSession().query(type).filter(
-        func.q3c_radial_query(type.ra,
-                              type.dec,
-                              self.ra, self.dec,
-                              0.64)
-    ).filter(
-        func.q3c_poly_query(self.ra, self.dec, type.poly)
-    )
+    altdata = sa.Column(psql.JSONB, nullable=True)
+    score = sa.Column(sa.Float, nullable=True)
+    accumulated_rb = sa.Column(sa.Float)
 
-    return candidates.all()
+    thumbnails = relationship('Thumbnail', cascade='all')
+    detections = relationship('Detection', cascade='all')
+    photometry = relationship('Photometry', back_populates='source',
+                              cascade='save-update, merge, refresh-expire, expunge',
+                              single_parent=True,
+                              passive_deletes=True,
+                              order_by="Photometry.observed_at")
 
-@without_database([])
-def best_detection(self):
-    return DBSession().query(
-        Detection
-    ).join(Source).filter(
-        Source.id == self.id
-    ).order_by(
-        (Detection.flux / Detection.fluxerr).desc()
-    ).first()
+    forced_photometry = relationship('ForcedPhotometry', cascade='all')
 
-Source.images = images
-Source.q3c = Index(
-    f'sources_q3c_ang2ipix_idx',
-    func.q3c_ang2ipix(
-        Source.ra,
-        Source.dec)
-)
+    def add_linked_thumbnails(self):
+        sdss_thumb = Thumbnail(photometry=self.photometry[0],
+                               public_url=self.sdss_url,
+                               type='sdss')
+        dr8_thumb = Thumbnail(photometry=self.photometry[0],
+                              public_url=self.desi_dr8_url,
+                              type='dr8')
+        DBSession().add_all([sdss_thumb, dr8_thumb])
+        DBSession().commit()
 
-Source.detections = relationship('Detection', cascade='all')
-Source.forced_photometry = relationship('ForcedPhotometry',
-                                               cascade='all')
-Source.best_detection = property(best_detection)
+    @property
+    def sdss_url(self):
+        """Construct URL for public Sloan Digital Sky Survey (SDSS) cutout."""
+        return (f"http://skyservice.pha.jhu.edu/DR9/ImgCutout/getjpeg.aspx"
+                f"?ra={self.ra}&dec={self.dec}&scale=0.3&width=200&height=200"
+                f"&opt=G&query=&Grid=on")
 
-Source.neighbor_info = sa.Column(psql.JSONB)
+    @property
+    def desi_dr8_url(self):
+        """Construct URL for public DESI DR8 cutout."""
+        return (f"http://legacysurvey.org/viewer/jpeg-cutout?ra={self.ra}"
+                f"&dec={self.dec}&size=200&layer=dr8&pixscale=0.262&bands=grz")
 
+    @without_database([])
+    def images(self, type=CalibratableImage):
 
-@without_database([])
-def unphotometered_images(self):
-    subq = DBSession().query(ForcedPhotometry.id,
-                             ForcedPhotometry.image_id).filter(
-        ForcedPhotometry.source_id == self.id
-    ).subquery()
-
-    q = DBSession().query(SingleEpochSubtraction).filter(
-        func.q3c_radial_query(SingleEpochSubtraction.ra,
-                              SingleEpochSubtraction.dec,
-                              self.ra, self.dec,
-                              0.64)
-    ).filter(
-        func.q3c_poly_query(self.ra, self.dec, SingleEpochSubtraction.poly)
-    ).outerjoin(
-        subq, subq.c.image_id == SingleEpochSubtraction.id
-    ).filter(
-        subq.c.id == None
-    )
-
-    return q
-
-
-def force_photometry(self, assume_background_subtracted=True):
-    out = []
-    for i in self.unphotometered_images:
-        sci_path = f'/global/cscratch1/sd/dgold/zuds/{i.field:06d}/' \
-                   f'c{i.ccdid:02d}/q{i.qid}/{fid_map[i.fid]}/' \
-                   f'{i.basename}'
-
-        mask_path = sci_path.replace('.fits', '.mask.fits')
-        rms_path = sci_path.replace('.fits', '.rms.fits')
-
-        fp = i.force_photometry(
-            self, assume_background_subtracted=assume_background_subtracted,
-            use_cutout=True, direct_load={'mask': mask_path, 'sci': sci_path,
-                                          'rms': rms_path}
+        candidates = DBSession().query(type).filter(
+            func.q3c_radial_query(type.ra,
+                                  type.dec,
+                                  self.ra, self.dec,
+                                  0.64)
+        ).filter(
+            func.q3c_poly_query(self.ra, self.dec, type.poly)
         )
 
-        out.extend(fp)
-    return out
+        return candidates.all()
 
+    @without_database([])
+    def best_detection(self):
+        return DBSession().query(
+            Detection
+        ).join(Source).filter(
+            Source.id == self.id
+        ).order_by(
+            (Detection.flux / Detection.fluxerr).desc()
+        ).first()
 
-Source.unphotometered_images = property(unphotometered_images)
-Source.force_photometry = force_photometry
+    @property
+    @without_database([])
+    def light_curve(self):
+        lc_raw = []
 
+        phot = DBSession().query(
+            ForcedPhotometry.obsjd - MJD_TO_JD,
+            ForcedPhotometry.filtercode,
+            ForcedPhotometry.zp,
+            ForcedPhotometry.flux,
+            ForcedPhotometry.fluxerr,
+            ForcedPhotometry.flags,
+            ForcedPhotometry.id
+        ).filter(
+            ForcedPhotometry.source_id == self.id
+        )
 
-@without_database([])
-def light_curve(sourceid):
-    lc_raw = []
+        for photpoint in phot:
+            photd = {'mjd': photpoint[0],
+                     'filter': 'ztf' + photpoint[1][-1],
+                     'zp': photpoint[2],  # ap-correct the ZP
+                     'zpsys': 'ab',
+                     'flux': photpoint[3],
+                     'fluxerr': photpoint[4],
+                     'flags': photpoint[5],
+                     'lim_mag': -2.5 * np.log10(5 * photpoint[4]) + photpoint[2],
+                     'id': photpoint[-1]}
+            lc_raw.append(photd)
 
-    phot = DBSession().query(
-        ForcedPhotometry.obsjd - MJD_TO_JD,
-        ForcedPhotometry.filtercode,
-        ForcedPhotometry.zp,
-        ForcedPhotometry.flux,
-        ForcedPhotometry.fluxerr,
-        ForcedPhotometry.flags,
-        ForcedPhotometry.id
-    ).filter(
-        ForcedPhotometry.source_id == sourceid
-    )
+        return Table(lc_raw)
 
-    for photpoint in phot:
-        photd = {'mjd': photpoint[0],
-                 'filter': 'ztf' + photpoint[1][-1],
-                 'zp': photpoint[2],  # ap-correct the ZP
-                 'zpsys': 'ab',
-                 'flux': photpoint[3],
-                 'fluxerr': photpoint[4],
-                 'flags': photpoint[5],
-                 'lim_mag': -2.5 * np.log10(5 * photpoint[4]) + photpoint[2],
-                 'id': photpoint[-1]}
-        lc_raw.append(photd)
+    @without_database([])
+    def unphotometered_images(self):
+        subq = DBSession().query(ForcedPhotometry.id,
+                                 ForcedPhotometry.image_id).filter(
+            ForcedPhotometry.source_id == self.id
+        ).subquery()
 
-    return Table(lc_raw)
+        q = DBSession().query(SingleEpochSubtraction).filter(
+            func.q3c_radial_query(SingleEpochSubtraction.ra,
+                                  SingleEpochSubtraction.dec,
+                                  self.ra, self.dec,
+                                  0.64)
+        ).filter(
+            func.q3c_poly_query(self.ra, self.dec, SingleEpochSubtraction.poly)
+        ).outerjoin(
+            subq, subq.c.image_id == SingleEpochSubtraction.id
+        ).filter(
+            subq.c.id == None
+        )
 
+        return q
 
-def lcprop(self):
-    return light_curve(self.id)
+    def force_photometry(self, assume_background_subtracted=True):
+        out = []
+        for i in self.unphotometered_images:
+            sci_path = f'/global/cscratch1/sd/dgold/zuds/{i.field:06d}/' \
+                       f'c{i.ccdid:02d}/q{i.qid}/{fid_map[i.fid]}/' \
+                       f'{i.basename}'
 
+            mask_path = sci_path.replace('.fits', '.mask.fits')
+            rms_path = sci_path.replace('.fits', '.rms.fits')
 
-Source.light_curve = lcprop
+            fp = i.force_photometry(
+                self, assume_background_subtracted=assume_background_subtracted,
+                use_cutout=True, direct_load={'mask': mask_path, 'sci': sci_path,
+                                              'rms': rms_path}
+            )
 
+            out.extend(fp)
+        return out
